seperate LoginViewController
